+++ src/activities/Elsa.Activities.Temporal.Common/Activities/ClearTimer/CancelTimerExtensions.cs
