+++ src/activities/Elsa.Activities.Temporal.Common/Activities/Timer/Timer.cs
