+++ src/activities/Elsa.Activities.Temporal.Common/Extensions/DurationExtensions.cs
