+++ src/activities/Elsa.Activities.Temporal.Common/FodyWeavers.xml
