+++ src/activities/Elsa.Activities.Temporal.Common/Handlers/RemoveScheduledTriggers.cs
