+++ src/activities/Elsa.Activities.Temporal.Common/Services/IWorkflowScheduler.cs
