+++ src/activities/Elsa.Activities.Temporal.Common/TimerConsts.cs
