+++ src/activities/Elsa.Activities.Timers.Quartz/Elsa.Activities.Timers.Quartz.csproj
@@ -19,7 +19,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\Elsa.Activities.Timers\Elsa.Activities.Timers.csproj" />
+      <ProjectReference Include="..\Elsa.Activities.Temporal.Common\Elsa.Activities.Temporal.Common.csproj" />
     </ItemGroup>
 
 </Project>
